deleted unnecessary files and added question api
