--- src/api/countrylanguagesmaster/index.js
@@ -1,8 +1,0 @@
-import { Router } from 'express';
-import { masterUpload } from './controller';
-
-const router = new Router();
-
-router.get('/', masterUpload);
-
-export default router;
